// src/lib/impakt-dev-api-client/react-query/types/getTimelineBlockResType.ts
/**
 * Generated by orval v6.10.2 🍺
 * Do not edit manually.
 * Impakt API
 * OpenAPI spec version: 1.0.0
 */

export type GetTimelineBlockResType =
  typeof GetTimelineBlockResType[keyof typeof GetTimelineBlockResType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetTimelineBlockResType = {
  Rest: 'Rest',
  HoldPose: 'HoldPose',
  HIIT: 'HIIT',
  Blitz: 'Blitz',
  Rhythm: 'Rhythm',
} as const;