// src/lib/impakt-dev-api-client/react-query/types/getTimelineBlockAttributeResType.ts
/**
 * Generated by orval v6.10.2 🍺
 * Do not edit manually.
 * Impakt API
 * OpenAPI spec version: 1.0.0
 */

export type GetTimelineBlockAttributeResType =
  typeof GetTimelineBlockAttributeResType[keyof typeof GetTimelineBlockAttributeResType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetTimelineBlockAttributeResType = {
  TimeConstraint: 'TimeConstraint',
  CountConstraint: 'CountConstraint',
  TimeGoal: 'TimeGoal',
  CountGoal: 'CountGoal',
} as const;